584% throughput increase

+++ src/com/vincentcodes/webserver/component/body/HttpBodyFileStream.java
@@ -14,31 +14,35 @@ import com.vincentcodes.webserver.component.header.EntityEncodings;
  * Does not apply any archiving, compression to the bytes 
  */
 public class HttpBodyFileStream implements HttpBody {
-    private File file;
-    private boolean deleteOnClose;
+    private final File file;
+    private final boolean deleteOnClose;
+    private final long lengthToRead;
+    private long readCount;
 
     private FileOutputStream fos;
     private FileInputStream fis;
-    private int maxCap = -1;
-    private int writtenCount = 0;
+    private long maxCap = -1;
+    private long writtenCount = 0;
     
-    public HttpBodyFileStream(File file, boolean deleteOnClose) throws FileNotFoundException{
+    public HttpBodyFileStream(File file, long lengthToRead, boolean deleteOnClose, boolean readonly) throws FileNotFoundException{
         this.file = file;
+        this.lengthToRead = lengthToRead;
         this.deleteOnClose = deleteOnClose;
         if(deleteOnClose) 
             file.deleteOnExit();
         
-        fos = new FileOutputStream(file);
+        if(!readonly)
+            fos = new FileOutputStream(file);
         fis = new FileInputStream(file);
     }
-    public HttpBodyFileStream(File file) throws FileNotFoundException{
-        this(file, false);
+    public HttpBodyFileStream(File file, long lengthToRead) throws FileNotFoundException{
+        this(file, lengthToRead, false, true);
     }
     /**
-     * Creates a temporary file to store and get the data
+     * Creates a temporary file to store and get the data. lengthToRead is set to 0
      */
     public HttpBodyFileStream() throws IOException{
-        this(File.createTempFile("webserver-res", null), true);
+        this(File.createTempFile("webserver-res", null), 0, true, false);
     }
 
     @Override
@@ -56,6 +60,8 @@ public class HttpBodyFileStream implements HttpBody {
 
     @Override
     public void writeToBody(int b) throws IOException{
+        if(fos == null) 
+            throw new IOException("This HttpBody is not configured to allow write operations on the file.");
         if(maxCap != -1){
             if(writtenCount > maxCap) return;
             writtenCount += 1;
@@ -70,6 +76,8 @@ public class HttpBodyFileStream implements HttpBody {
     
     @Override
     public void writeToBody(byte[] b, int length) throws IOException {
+        if(fos == null) 
+            throw new IOException("This HttpBody is not configured to allow write operations on the file.");
         if(maxCap != -1){
             if(writtenCount > maxCap) return;
             writtenCount += length;
@@ -102,7 +110,11 @@ public class HttpBodyFileStream implements HttpBody {
      */
     @Override
     public byte[] getBytes(){
+        if(readCount >= length()){
+            return new byte[0];
+        }
         try{
+            readCount += length();
             return fis.readAllBytes();
         }catch(IOException e){
             e.printStackTrace();
@@ -117,14 +129,39 @@ public class HttpBodyFileStream implements HttpBody {
      */
     @Override
     public byte[] getBytes(int length) {
+        if(readCount >= length()){
+            return new byte[0];
+        }
         try{
-            return fis.readNBytes(length);
+            readCount += length;
+            return fis.readNBytes(readCount > length()? (int)(length() - (readCount - length)) : length);
         }catch(IOException e){
             e.printStackTrace();
             return new byte[0];
         }
     }
 
+    /**
+     * @return -1 while error or eof is reached
+     */
+    @Override
+    public int getBytes(byte[] buffer){
+        if(readCount >= length()){
+            return -1;
+        }
+        try{
+            int bytesRead = fis.read(buffer);
+            readCount += bytesRead; 
+            return bytesRead;
+        }catch(IOException e){
+            e.printStackTrace();
+            return -1;
+        }
+    }
+
+    /**
+     * length will not apply here.
+     */
     @Override
     public void streamBytesTo(OutputStream os) throws IOException {
         fis.transferTo(os);
@@ -134,20 +171,16 @@ public class HttpBodyFileStream implements HttpBody {
      * @return available bytes up for grab
      */
     @Override
-    public int length(){
-        try {
-            return fis.available();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return 0;
+    public long length(){
+        return lengthToRead;
     }
 
     @Override
     public void close() throws IOException {
         try{
             fis.close();
-            fos.close();
+            if(fos != null)
+                fos.close();
         }catch(IOException e){
             e.printStackTrace();
         }finally{
@@ -157,10 +190,19 @@ public class HttpBodyFileStream implements HttpBody {
     }
 
     /**
+     * Skip number of bytes for FileInputStream
+     * @throws IOException
+     */
+    public void skip(long numOfBytes) throws IOException{
+        fis.skip(numOfBytes);
+    }
+
+    /**
      * When error occurs, input stream will NOT be reset.
      */
     public void resetInput() throws IOException{
         fis.close();
         fis = new FileInputStream(file);
+        readCount = 0;
     }
 }
